% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pd_presslog.R
\name{get_pd_presslog}
\alias{get_pd_presslog}
\title{Download the current presslog}
\usage{
get_pd_presslog(
  url = "https://data.city.ames.ia.us/publicinformation/PressLog.pdf",
  save_as = sprintf("presslog_amespd_\%s.pdf", lubridate::today())
)
}
\arguments{
\item{url}{url to download a press log from}

\item{save_as}{file path to save press log to}
}
\value{
Character of the location of the saved pdf file.
}
\description{
Download the current presslog
}
