% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presslog.R
\name{get_presslog_ames}
\alias{get_presslog_ames}
\title{Get the presslog for Ames PD for an interval}
\usage{
get_presslog_ames(from = lubridate::today(), to = lubridate::today())
}
\arguments{
\item{from}{date in YYYY-MM-DD format}

\item{to}{date in YYYY-MM-DD format}
}
\value{
data frame with call logs
}
\description{
Get the presslog for Ames PD for an interval
}
