% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_tbl.R
\name{pdf_to_tbl}
\alias{pdf_to_tbl}
\title{Convert pdf presslog into a csv table}
\usage{
pdf_to_tbl(pdfs)
}
\arguments{
\item{pdfs}{path(s) to a (set of) pdf file(s)}
}
\description{
Convert pdf presslog into a csv table
}
