% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presslog_ames}
\alias{presslog_ames}
\title{The Ames Police Department Press Log data set}
\format{
A data frame of past Ames Police Department Press Logs.
\describe{
\item{Call Received Date/Time}{when the service call was received}
\item{Incident ID}{incident id}
\item{How Call was Rec'd}{the medium used to contact Ames PD}
\item{Nature Code Description}{description of the incident}
\item{Location Address}{where the incident occured}
\item{Report Number Assigned to Event}{report number if a report was generated}
\item{Closing Disposition or Cancel Code}{short acronym describing the type of incident}
}
}
\source{
Ames Police Department
}
\usage{
presslog_ames
}
\description{
The Ames Police Department Press Log data set
}
\examples{
head(presslog_ames)
dim(presslog_ames)
}
\keyword{datasets}
