% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presslog_isu}
\alias{presslog_isu}
\title{The ISU Police Department Press Log data set}
\format{
A data frame of past ISU Police  Press Logs.
\describe{
\item{Case.Number}{case number}
\item{Date.Time.Reported}{date of report}
\item{Earliest.Occurrence}{date of earliest occurrence}
\item{Latest.Occurrence}{date of latest occurrence}
\item{General.Location}{description of the location}
\item{Disposition}{is it closed?}
\item{Classification}{on the report. so many ways to spell things}
\item{longitude}{in degree}
\item{latitude}{in degree}
}
}
\source{
ISU Police Department
}
\usage{
presslog_isu
}
\description{
The ISU Police Department Press Log data set
}
\examples{
head(presslog_isu)
dim(presslog_isu)
}
\keyword{datasets}
